+++ src/utils/clack-utils.ts
@@ -21,6 +21,8 @@ import {
 import { debug } from './debug';
 import { fulfillsVersionRange } from './semver';
 
+import { matchesContent } from '../../lib/Helper/File';
+
 const opn = require('opn') as (
   url: string,
 ) => Promise<childProcess.ChildProcess>;
@@ -614,7 +616,16 @@ SENTRY_AUTH_TOKEN=${authToken}
 }
 
 async function addCliConfigFileToGitIgnore(filename: string): Promise<void> {
-  //TODO: Add a check to see if the file is already ignored in .gitignore
+  const regex = new RegExp(`^\\s*${filename}\\s*`, 'gm');
+  if (matchesContent(filename, regex)) {
+    clack.log.info(
+      `${chalk.bold('.gitignore')} already has ${chalk.bold(
+        filename,
+      )}. Will not add it again.`,
+    );
+    return;
+  }
+
   try {
     await fs.promises.appendFile(
       path.join(process.cwd(), '.gitignore'),